// ancore-gateway/src/types/error.ts
export enum ErrorDefs {
  NOT_FOUND = 'NOT_FOUND',
  MISING_FIELD = 'MISING_FIELD',
  MISING_DATA = 'MISING_DATA',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
  UNAUTHORIZED = 'UNAUTHORIZED',
  UNAUTHENTICATED = 'UNAUTHENTICATED',
  INVALID_TOKEN = 'INVALID_TOKEN',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  INVALID_REQUEST = 'INVALID_REQUEST',
  INVALID_INPUT = 'INVALID_INPUT',
  INVALID_FILE = 'INVALID_FILE',
  BAD_REQUEST = 'BAD_REQUEST',
  FORBIDDEN= 'FORBIDDEN',
  ALREADY_EXISTS = 'ALREADY_EXISTS'
}